Uses shared helpers for github lookups
